seagul/old/sym_rl/requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
dill==0.2.9
fasteners==0.15
future==0.18.2
glfw==1.8.4
grpcio==1.14.1
gym==0.15.3
idna==2.8
imageio==2.6.1
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
joblib==0.13.2
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
lockfile==0.12.2
matplotlib==3.1.1
monotonic==1.5
numpy==1.17.3
opencv-contrib-python==4.0.0.21
opencv-python==4.1.0.25
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.2.0
pixiedust==1.1.15
progressbar2==3.39.3
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pybullet==2.5.6
pycparser==2.19
pyglet==1.3.2
Pygments==2.4.2
PyOpenGL==3.1.0
pyparsing==2.4.4
python-dateutil==2.8.1
python-utils==2.3.0
pyzmq==18.1.0
requests==2.21.0
scikit-learn==0.21.3
scipy==1.3.1
simplegeneric==0.8.1
six==1.13.0
sklearn==0.0
termcolor==1.1.0
torchfile==0.1.0
tornado==6.0.3
tqdm==4.31.1
traitlets==4.3.3
urllib3==1.24.1
wcwidth==0.1.7
zmq==0.0.0
